scraper/requirements.txt
requests==2.31.0
pandas==2.1.3
supabase==2.0.3
python-dotenv==1.0.0